update commands only on argument

+++ src/main/java/de/janno/discord/command/SlashCommandRegistry.java
@@ -37,50 +37,52 @@ public class SlashCommandRegistry {
             return this;
         }
 
-        public SlashCommandRegistry registerSlashCommands(DiscordClient discordClient) {
-            ApplicationService applicationService = discordClient.getApplicationService();
-            long applicationId = discordClient.getApplicationId().blockOptional().orElseThrow();
-            Map<String, ISlashCommand> buttonSlashCommandMap = slashCommands.stream()
-                    .collect(Collectors.toMap(ISlashCommand::getName, Function.identity()));
-            log.info("ApplicationId: " + applicationId);
+        public SlashCommandRegistry registerSlashCommands(DiscordClient discordClient, boolean update) {
+            if (update) {
+                ApplicationService applicationService = discordClient.getApplicationService();
+                long applicationId = discordClient.getApplicationId().blockOptional().orElseThrow();
+                Map<String, ISlashCommand> buttonSlashCommandMap = slashCommands.stream()
+                        .collect(Collectors.toMap(ISlashCommand::getName, Function.identity()));
+                log.info("ApplicationId: " + applicationId);
 
-            //delete old commands
-            applicationService.getGlobalApplicationCommands(applicationId)
-                    .doOnEach(acd -> log.info("Existing command: " + acd))
-                    .filter(acd -> !buttonSlashCommandMap.containsKey(acd.name()))
-                    .doOnEach(acd -> log.info("Deleting command: " + acd))
-                    .flatMap(acd -> applicationService.deleteGlobalApplicationCommand(applicationId, Long.parseLong(acd.id())))
-                    .subscribe();
+                //delete old commands
+                applicationService.getGlobalApplicationCommands(applicationId)
+                        .doOnEach(acd -> log.info("Existing command: " + acd))
+                        .filter(acd -> !buttonSlashCommandMap.containsKey(acd.name()))
+                        .doOnEach(acd -> log.info("Deleting command: " + acd))
+                        .flatMap(acd -> applicationService.deleteGlobalApplicationCommand(applicationId, Long.parseLong(acd.id())))
+                        .subscribe();
 
-            //get already existing commands
-            Map<String, ApplicationCommandData> applicationCommandData = applicationService.getGlobalApplicationCommands(applicationId)
-                    .collectMap(ApplicationCommandData::name, Function.identity()).blockOptional().orElse(ImmutableMap.of());
+                //get already existing commands
+                Map<String, ApplicationCommandData> applicationCommandData = applicationService.getGlobalApplicationCommands(applicationId)
+                        .collectMap(ApplicationCommandData::name, Function.identity()).blockOptional().orElse(ImmutableMap.of());
 
-            //add missing commands
-            buttonSlashCommandMap.values().stream()
-                    .filter(sc -> !applicationCommandData.containsKey(sc.getName()))
-                    .forEach(sc -> {
-                                log.info("Creating command: " + sc.getApplicationCommand());
-                                applicationService.createGlobalApplicationCommand(applicationId, sc.getApplicationCommand())
-                                        .doOnEach(acd -> log.info("Created command: " + acd))
-                                        .block();
-                            }
-                    );
+                //add missing commands
+                buttonSlashCommandMap.values().stream()
+                        .filter(sc -> !applicationCommandData.containsKey(sc.getName()))
+                        .forEach(sc -> {
+                                    log.info("Creating command: " + sc.getApplicationCommand());
+                                    applicationService.createGlobalApplicationCommand(applicationId, sc.getApplicationCommand())
+                                            .doOnEach(acd -> log.info("Created command: " + acd))
+                                            .block();
+                                }
+                        );
 
-            //update existing
-            //todo check for changes in existing commands
+                //update existing
+                //todo check for changes in existing commands
 
-            buttonSlashCommandMap.values().stream()
-                    .filter(sc -> applicationCommandData.containsKey(sc.getName()))
-                    .forEach(sc -> {
-                                log.info("Creating command: " + sc.getApplicationCommand());
-                                applicationService.modifyGlobalApplicationCommand(applicationId,
-                                        Long.parseLong(applicationCommandData.get(sc.getName()).id()),
-                                        sc.getApplicationCommand())
-                                        .doOnEach(acd -> log.info("Update command: " + acd))
-                                        .block();
-                            }
-                    );
+                buttonSlashCommandMap.values().stream()
+                        .filter(sc -> applicationCommandData.containsKey(sc.getName()))
+                        .forEach(sc -> {
+                                    log.info("Creating command: " + sc.getApplicationCommand());
+                                    applicationService.modifyGlobalApplicationCommand(applicationId,
+                                            Long.parseLong(applicationCommandData.get(sc.getName()).id()),
+                                            sc.getApplicationCommand())
+                                            .doOnEach(acd -> log.info("Update command: " + acd))
+                                            .block();
+                                }
+                        );
+            }
             return new SlashCommandRegistry(slashCommands);
         }
